Move vars to std namespace
